pool_operation_mappings {} calculate_total_denorm_weight() moved
